2022.05.02 21:24 Update

+++ Array/0034. Find First and Last Position of Element in Sorted Array.cpp
@@ -4,6 +4,6 @@ public:
 		auto it = lower_bound(nums.begin(), nums.end(), target);
 		if (it == nums.end() || *it != target)
 			return { -1, -1 };
-		return { it - nums.begin(), upper_bound(nums.begin(), nums.end(), target) - nums.begin() - 1 };
+		return { static_cast<int>(it - nums.begin()), static_cast<int>(upper_bound(nums.begin(), nums.end(), target) - nums.begin()) - 1 };
 	}
 };
